Rename PersistentEventHistory to FileEventHistory
We might have a DB-backed history in the future.
